solve_schrodinger_gpu now is solve_schrodinger with multiple dispatch
Before, we had a separate solve_schrodinger function for GPU. Now, the
single function solve_schrodinger can be called to use GPU or not by
inferring from the type of A, the annealing object.

If A::Annealing --> use CPU
IF A::CuAnnealing --> use GPU

This requires the user to define a CuAnnealing object instead of Annealing
if they want to use a GPU. We think this is fair since certain GPU
optimizations should be made even in the constructors. This also
makes the solvers have a cleaner interface (no additional flag--just infers
from argument types).